Quick search in the swap drop-down
