+++ assets/css/app.scss
@@ -78,5 +78,5 @@ $font-size-base: 0.95rem !default;
 
 .dropdown-menu {
   max-height: 500px;
-  overflow-y: scroll;
+  overflow-y: auto;
 }
