+++ mittab/apps/tab/templates/judge_dropdown.html
@@ -1,3 +1,8 @@
+<div class="form-group pl-2 pt-2 mb-1">
+  <input id="quick-search"
+        class="form-control form-control-sm" type="text" placeholder="Search...">
+</div>
+
 <h6 class="text-center dropdown-header">Current Judge</h6>
 
 <a class="dropdown-item "href="/judge/{{current_judge_id}}">
